# Conflicts:
#	components/asic/asic.c
#	components/asic/bm1370.c
#	components/asic/include/asic.h
#	components/asic/include/bm1366.h
#	components/asic/include/bm1368.h
#	components/asic/include/bm1370.h
#	components/asic/include/bm1397.h
#	components/connect/connect.c
#	components/connect/include/connect.h
#	main/adc.c
#	main/global_state.h
#	main/http_server/axe-os/api/system/asic_settings.c
#	main/http_server/axe-os/package-lock.json
#	main/http_server/axe-os/package.json
#	main/http_server/axe-os/src/app/components/home/home.component.html
#	main/http_server/axe-os/src/app/components/home/home.component.ts
#	main/http_server/axe-os/src/app/components/settings/settings.component.html
#	main/http_server/axe-os/src/app/components/settings/settings.component.ts
#	main/http_server/http_server.c
#	main/main.c
#	main/main.h
#	main/nvs_config.h
#	main/power/TPS546.h
#	main/screen.c
#	main/self_test/self_test.c
#	main/self_test/self_test.h
#	main/system.c
#	main/system.h
#	main/tasks/asic_result_task.c
#	main/tasks/create_jobs_task.c
#	main/tasks/power_management_task.c
#	main/tasks/statistics_task.c
#	main/tasks/stratum_task.c
#	main/thermal/thermal.c

+++ main/adc.c
@@ -104,5 +104,5 @@ uint16_t ADC_get_vcore(void)
         return 0;
     }
 
-    return (uint16_t)voltage;
+    return voltage;
 }
